// generated/orval/fetch/google-health-api/models/settingsDistanceUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. The measurement unit defined in the user's account settings.
 */
export type SettingsDistanceUnit = typeof SettingsDistanceUnit[keyof typeof SettingsDistanceUnit];


export const SettingsDistanceUnit = {
  DISTANCE_UNIT_UNSPECIFIED: 'DISTANCE_UNIT_UNSPECIFIED',
  DISTANCE_UNIT_MILES: 'DISTANCE_UNIT_MILES',
  DISTANCE_UNIT_KILOMETERS: 'DISTANCE_UNIT_KILOMETERS',
} as const;