// generated/orval/fetch/google-health-api/models/settingsSwimUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. The measurement unit defined in the user's account settings.
 */
export type SettingsSwimUnit = typeof SettingsSwimUnit[keyof typeof SettingsSwimUnit];


export const SettingsSwimUnit = {
  SWIM_UNIT_UNSPECIFIED: 'SWIM_UNIT_UNSPECIFIED',
  SWIM_UNIT_METERS: 'SWIM_UNIT_METERS',
  SWIM_UNIT_YARDS: 'SWIM_UNIT_YARDS',
} as const;