// generated/orval/fetch/google-health-api/models/settingsTemperatureUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. The measurement unit defined in the user's account settings.
 */
export type SettingsTemperatureUnit = typeof SettingsTemperatureUnit[keyof typeof SettingsTemperatureUnit];


export const SettingsTemperatureUnit = {
  TEMPERATURE_UNIT_UNSPECIFIED: 'TEMPERATURE_UNIT_UNSPECIFIED',
  TEMPERATURE_UNIT_CELSIUS: 'TEMPERATURE_UNIT_CELSIUS',
  TEMPERATURE_UNIT_FAHRENHEIT: 'TEMPERATURE_UNIT_FAHRENHEIT',
} as const;