// generated/orval/fetch/google-health-api/models/settingsWaterUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. The measurement unit defined in the user's account settings.
 */
export type SettingsWaterUnit = typeof SettingsWaterUnit[keyof typeof SettingsWaterUnit];


export const SettingsWaterUnit = {
  WATER_UNIT_UNSPECIFIED: 'WATER_UNIT_UNSPECIFIED',
  WATER_UNIT_ML: 'WATER_UNIT_ML',
  WATER_UNIT_FL_OZ: 'WATER_UNIT_FL_OZ',
  WATER_UNIT_CUP: 'WATER_UNIT_CUP',
} as const;