// generated/orval/fetch/google-health-api/models/settingsWeightUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. The measurement unit defined in the user's account settings.
 */
export type SettingsWeightUnit = typeof SettingsWeightUnit[keyof typeof SettingsWeightUnit];


export const SettingsWeightUnit = {
  WEIGHT_UNIT_UNSPECIFIED: 'WEIGHT_UNIT_UNSPECIFIED',
  WEIGHT_UNIT_POUNDS: 'WEIGHT_UNIT_POUNDS',
  WEIGHT_UNIT_STONE: 'WEIGHT_UNIT_STONE',
  WEIGHT_UNIT_KILOGRAMS: 'WEIGHT_UNIT_KILOGRAMS',
} as const;